Remove tst_qtconcurrentiteratekernel blockSize test
The test is just wrong.

It suggests, that it tests minimal block size,
but in reality it is checking if a maximal peak of block size
is bigger then expected minimum block size.

The minimal block size is 1, not 1024 / ideal threads count and
it is defined in BlockSizeManagerV2.

The maximal block size is defined and it is fixed, but it is an
implementation detail, probably not worth testing.

The test is based on a race condition as peakBegin and
peakBlockSize are modified in parallel from multiple threads,
without any synchronization.

Change-Id: I430eedcf83b0fa3e2ce2cfd294eaee7b301e48ae
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>