Windows/QFileSystemModel: Fix updating of removed drives
Previously, the updating of drives in QFileSystemModel was connected
to a signal triggering when a drive containing watched files was
removed via QFileSystemWatcher notification. This did not trigger
when a drive that was not expanded in the view was removed, since
no files were watched.
Since QFileSystemModel is not interested in the path of the drive
being removed, add a generic signal triggered by
DBT_DEVTYP_VOLUME/DBT_DEVICEREMOVECOMPLETE and use that to update
the drives.
Complements 8e79806d08ab77aa0f87b69a2ef65789216f41c0.

Task-number: QTBUG-18729
Task-number: QTBUG-53436
Change-Id: Ibcde4665824c41151042237d4d620c48bc1e2e18
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>