eglfs_kms: Add support for formats other than XRGB8888
Switch to drmModeAddFB2 and add mappings for RGB565 and ARGB8888,
and some BGR variants.

The default is XRGB8888, like before.

Others can be requested in the config file with
"format": "rgb565" or "argb8888" etc. on a per-output basis.

If the primary plane does not support the format, modesetting
and flipping will fail.

Task-number: QTBUG-62262
Change-Id: I8537cbeed7f046c46aa63fcea6d6946c0e0038a7
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>