Android: Fix push button label duplication with stylesheets
In the Android style, we matched both the full CE_PushButton,
and the individual, decomposed CE_PushButtonBevel and
CE_PushButtonLabel (and similar for other controls) to the
same drawing code which draws both the label and the bevel.
When used together with the style sheet style, this caused
the label of the button to be drawn multiple times in
different locations.

For labels, we now skip the part which draws the actual control,
and for the non-textual parts of the controls, we skip drawing
the label.

[ChangeLog][QtWidgets][Android] Fixed label duplication for
buttons when using style sheets with the Android style.

Task-number: QTBUG-48639
Change-Id: I1a1cdb0a82870cebd0438c5c72199afc3192938c
Reviewed-by: BogDan Vatra <bogdan@kdab.com>